refactor: moved dequeueing init process to a separate class
