+++ entities-viewings-collector/src/main/scala/io/renku/entities/viewings/collector/projects/viewed/SubscriptionFactory.scala
@@ -18,7 +18,7 @@
 
 package io.renku.entities.viewings.collector.projects.viewed
 
-import cats.effect.{Async, Temporal}
+import cats.effect.Async
 import cats.syntax.all._
 import io.renku.db.SessionResource
 import io.renku.events.consumers.subscriptions.SubscriptionMechanism
@@ -27,8 +27,6 @@ import io.renku.eventsqueue.EventsQueue
 import io.renku.lock.Lock
 import io.renku.triplesstore.{ProjectsConnectionConfig, SparqlQueryTimeRecorder}
 import org.typelevel.log4cats.Logger
-
-import scala.concurrent.duration._
 
 object SubscriptionFactory {
   def apply[F[_]: Async: Logger: SparqlQueryTimeRecorder, DB](
@@ -42,50 +40,7 @@ object SubscriptionFactory {
       eventsQueue    <- EventsQueue[F, DB](categoryLock).pure[F]
       handler        <- EventHandler[F](eventsQueue)
       eventProcessor <- EventProcessor[F](connConfig, eventsQueue)
-      _              <- kickOffEventsDequeueing[F](tsReadyCheck, eventsQueue, eventProcessor)
+      _              <- EventsDequeueingInitiator(tsReadyCheck, eventsQueue, eventProcessor).initDequeueing
     } yield handler -> SubscriptionMechanism.noOpSubscriptionMechanism(categoryName)
   }
-
-  private[viewed] def kickOffEventsDequeueing[F[_]: Async: Logger](tsReadyCheck:     F[Boolean],
-                                                                   eventsQueue:      EventsQueue[F],
-                                                                   processor:        EventProcessor[F],
-                                                                   onErrorWait:      Duration = 2 seconds,
-                                                                   onTSNotReadyWait: Duration = 10 seconds
-  ) = Async[F].start {
-    startDequeueingEvents(tsReadyCheck, eventsQueue, processor, onErrorWait, onTSNotReadyWait)
-  }.void
-
-  private def startDequeueingEvents[F[_]: Async: Logger](tsReadyCheck:     F[Boolean],
-                                                         eventsQueue:      EventsQueue[F],
-                                                         processor:        EventProcessor[F],
-                                                         onErrorWait:      Duration,
-                                                         onTSNotReadyWait: Duration
-  ): F[Unit] = tsReadyCheck
-    .flatMap {
-      case true =>
-        Logger[F].info(show"Starting events dequeueing for $categoryName") >>
-          hookToTheEventsStream(eventsQueue, processor)
-      case false =>
-        Logger[F].info(show"TS not ready for writing; dequeueing for $categoryName on hold") >>
-          Temporal[F].delayBy(
-            startDequeueingEvents(tsReadyCheck, eventsQueue, processor, onErrorWait, onTSNotReadyWait),
-            onTSNotReadyWait
-          )
-    }
-    .handleErrorWith(waitAndRetry(tsReadyCheck, eventsQueue, processor, onErrorWait, onTSNotReadyWait))
-
-  private def waitAndRetry[F[_]: Async: Logger](tsReadyCheck:     F[Boolean],
-                                                eventsQueue:      EventsQueue[F],
-                                                processor:        EventProcessor[F],
-                                                onErrorWait:      Duration,
-                                                onTSNotReadyWait: Duration
-  ): Throwable => F[Unit] =
-    Logger[F].error(_)(show"An error in the $categoryName processing pipe; restarting") >>
-      Temporal[F].delayBy(
-        startDequeueingEvents(tsReadyCheck, eventsQueue, processor, onErrorWait, onTSNotReadyWait),
-        onErrorWait
-      )
-
-  private def hookToTheEventsStream[F[_]: Async](eventsQueue: EventsQueue[F], processor: EventProcessor[F]) =
-    eventsQueue.acquireEventsStream(categoryName).through(processor).compile.drain
 }
